+++ src/main/java/net/thedragonteam/armorplus/api/crafting/hightechbench/recipes/ModItemRecipes.java
@@ -17,7 +17,7 @@ public class ModItemRecipes {
                 "LUUL",
                 "OWHO",
                 "OUUO",
-                "OXXO",
+                "O  O",
                 'U', ModItems.theUltimateMaterial,
                 'W', ModBlocks.arpWorkbench,
                 'H', ModBlocks.arpHighTechBench,