+++ src/main/java/net/thedragonteam/armorplus/blocks/castle/RedStoneBrick.java
@@ -9,24 +9,17 @@ import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.renderer.block.model.ModelResourceLocation;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemBlock;
-import net.minecraft.util.BlockRenderLayer;
-import net.minecraft.util.EnumBlockRenderType;
 import net.minecraftforge.client.model.ModelLoader;
 import net.minecraftforge.fml.common.registry.GameRegistry;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 import net.thedragonteam.armorplus.blocks.castle.base.BaseCastleBlock;
 
-/**
- * net.thedragonteam.armorplus.blocks
- * ArmorPlus created by sokratis12GR on 6/13/2016 9:46 PM.
- * - TheDragonTeam
- */
-public class WhiteStoneBrickTower extends BaseCastleBlock {
+public class RedStoneBrick extends BaseCastleBlock {
 
-    public WhiteStoneBrickTower() {
-        super("white_stone_brick_tower");
-        setRegistryName("white_stone_brick_tower");
+    public RedStoneBrick() {
+        super("red_stone_brick");
+        setRegistryName("red_stone_brick");
         GameRegistry.register(this);
         GameRegistry.register(new ItemBlock(this), getRegistryName());
     }
@@ -40,27 +33,6 @@ public class WhiteStoneBrickTower extends BaseCastleBlock {
      * Get the MapColor for this Block and the given BlockState
      */
     public MapColor getMapColor(IBlockState state) {
-        return MapColor.QUARTZ;
-    }
-
-    @Override
-    public boolean isOpaqueCube(IBlockState state) {
-        return false;
-    }
-
-    @Override
-    public boolean isFullCube(IBlockState state) {
-        return false;
-    }
-
-    @Override
-    public EnumBlockRenderType getRenderType(IBlockState state) {
-        return EnumBlockRenderType.MODEL;
-    }
-
-    @Override
-    @SideOnly(Side.CLIENT)
-    public BlockRenderLayer getBlockLayer() {
-        return BlockRenderLayer.CUTOUT;
+        return MapColor.RED;
     }
 }