+++ src/main/java/net/thedragonteam/armorplus/blocks/castle/WhiteStoneBrick.java
@@ -4,6 +4,8 @@
 
 package net.thedragonteam.armorplus.blocks.castle;
 
+import net.minecraft.block.material.MapColor;
+import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.renderer.block.model.ModelResourceLocation;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemBlock;
@@ -32,4 +34,10 @@ public class WhiteStoneBrick extends BaseCastleBlock {
         ModelLoader.setCustomModelResourceLocation(Item.getItemFromBlock(this), 0, new ModelResourceLocation(getRegistryName(), "inventory"));
     }
 
+    /**
+     * Get the MapColor for this Block and the given BlockState
+     */
+    public MapColor getMapColor(IBlockState state) {
+        return MapColor.QUARTZ;
+    }
 }