+++ src/main/java/net/thedragonteam/armorplus/blocks/castle/base/BaseCastleBlock.java
@@ -5,9 +5,7 @@
 package net.thedragonteam.armorplus.blocks.castle.base;
 
 import net.minecraft.block.Block;
-import net.minecraft.block.material.MapColor;
 import net.minecraft.block.material.Material;
-import net.minecraft.block.state.IBlockState;
 import net.thedragonteam.armorplus.ArmorPlus;
 
 public class BaseCastleBlock extends Block {
@@ -19,12 +17,5 @@ public class BaseCastleBlock extends Block {
         this.setCreativeTab(ArmorPlus.tabArmorplusBlocks);
         this.setHardness(5.0F);
         this.setHarvestLevel("pickaxe", 0);
-    }
-
-    /**
-     * Get the MapColor for this Block and the given BlockState
-     */
-    public MapColor getMapColor(IBlockState state) {
-        return MapColor.QUARTZ;
     }
 }