+++ src/main/java/net/thedragonteam/armorplus/items/bows/EmeraldBow.java
@@ -6,7 +6,6 @@ package net.thedragonteam.armorplus.items.bows;
 
 import net.minecraft.init.Blocks;
 import net.minecraft.init.Items;
-import net.minecraft.item.Item;
 import net.minecraft.util.text.TextFormatting;
 import net.thedragonteam.armorplus.items.base.BaseBow;
 import net.thedragonteam.armorplus.registry.ModItems;
@@ -19,6 +18,6 @@ import net.thedragonteam.armorplus.registry.ModItems;
 public class EmeraldBow extends BaseBow {
 
     public EmeraldBow() {
-        super(1561, "emerald_bow", 5.0F, Items.EMERALD, Item.getItemFromBlock(Blocks.EMERALD_BLOCK), TextFormatting.DARK_GREEN, ModItems.emeraldBow);
+        super(1561, "emerald_bow", 5.0F, Items.EMERALD, Blocks.EMERALD_BLOCK, TextFormatting.DARK_GREEN, ModItems.emeraldBow);
     }
 }